+++ io/src/main/java/com/github/jorgecastillo/kotlinandroid/io/algebras/business/HeroesUseCases.kt
@@ -1,10 +1,10 @@
-package com.github.jorgecastillo.kotlinandroid.io.algebras.services
+package com.github.jorgecastillo.kotlinandroid.io.algebras.business
 
 import arrow.effects.IO
 import com.github.jorgecastillo.kotlinandroid.io.algebras.persistence.HeroesRepository
 import com.karumi.marvelapiclient.model.CharacterDto
 
-object HeroesService {
+object HeroesUseCases {
 
   fun getHeroes(): IO<List<CharacterDto>> =
       HeroesRepository.getHeroesWithCachePolicy(
