+++ io/src/main/java/com/github/jorgecastillo/kotlinandroid/io/algebras/business/model/SuperHero.kt
@@ -1,0 +1,3 @@
+package com.github.jorgecastillo.kotlinandroid.io.algebras.business.model
+
+data class SuperHero(val name: String)
