--- io/src/main/java/com/github/jorgecastillo/kotlinandroid/io/algebras/services/model/SuperHero.kt
@@ -1,3 +1,0 @@
-package com.github.jorgecastillo.kotlinandroid.io.algebras.services.model
-
-data class SuperHero(val name: String)
